Manager fix

+++ src/main/java/sg/nus/iss/team2/service/LeaveBalanceService.java
@@ -24,7 +24,7 @@ public interface LeaveBalanceService {
 
     void minusMedicalLeaveBalance(LeaveBalance leaveBalance, int days);
 
-    void minusCompensationLeaveBalance(LeaveBalance leaveBalance, int days);
+    void minusCompensationLeaveBalance(LeaveBalance leaveBalance, double days);
 
     // After Manager approve compensation request. Increase compensation leave balance.
     void addCompensationLeaveBalance(LeaveBalance leaveBalance, double days);
